scripts/fastapiforge/examples/requirements.txt
fastapi[standard]
pydantic
uvicorn
sqlmodel
psycopg2-binary
sqlalchemy[asyncio]
python-dotenv